update files from js to jsx
